# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  posts.CreatePostRequest:
    properties:
      caption:
        maxLength: 1000
        type: string
      image_url:
        description: Can be file_key from MinIO or full URL
        type: string
    required:
    - caption
    - image_url
    type: object
  posts.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  posts.Post:
    properties:
      caption:
        type: string
      created_at:
        type: string
      image_url:
        type: string
      post_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  posts.PostResponse:
    properties:
      data:
        $ref: '#/definitions/posts.Post'
      message:
        type: string
      success:
        type: boolean
    type: object
  posts.UpdatePostRequest:
    properties:
      caption:
        maxLength: 1000
        type: string
      image_url:
        description: Can be file_key from MinIO or full URL
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@instant.com
    name: API Support
  description: |-
    Microservices API Gateway for Instant social media platform
    This API provides access to authentication, posts, files, likes, comments, and follow services.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Instant API Gateway
  version: "1.0"
paths:
...
  /api/posts:
    get:
      description: Retrieve all posts with pagination support
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      summary: Get all posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a new post with caption and image URL (requires authentication)
      parameters:
      - description: Post creation data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/posts.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Create a new post
      tags:
      - posts
  /api/posts/{id}:
    delete:
      description: Delete a post by ID (requires authentication and ownership)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      description: Retrieve a single post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      summary: Get a post by ID
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Update post caption and/or image URL (requires authentication and
        ownership)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Post update data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/posts.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/posts.PostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Update a post
      tags:
      - posts
...
  /api/users/{user_id}/posts:
    get:
      description: Retrieve all posts created by a specific user with pagination
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/posts.ErrorResponse'
      summary: Get posts by user
      tags:
      - posts
...
securityDefinitions:
  SessionAuth:
    description: Session-based authentication via cookie
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"